+++ test/srt/models/test_vlm_models.py
@@ -5,7 +5,7 @@ import unittest
 from types import SimpleNamespace
 
 from sglang.srt.utils import is_hip
-from sglang.test.kits.mmmu_vlm_mixin import DEFAULT_MEM_FRACTION_STATIC, MMMUVLMMixin
+from sglang.test.kits.mmmu_vlm_kit import DEFAULT_MEM_FRACTION_STATIC, MMMUVLMMixin
 from sglang.test.test_utils import CustomTestCase, is_in_ci
 
 _is_hip = is_hip()
